Fix for private buy.
git-svn-id: http://svn.l2jserver.com/trunk@3814 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/com/l2jserver/gameserver/network/clientpackets/SetPrivateStoreListBuy.java
@@ -115,7 +115,7 @@ public final class SetPrivateStoreListBuy extends L2GameClientPacket
 			return;
 		}
 
-		int totalCost = 0;
+		long totalCost = 0;
 		for (Item i : _items)
 		{
 			if (!i.addToTradeList(tradeList))
